optimized minimap

+++ src/render_next_frame/show_map2d.c
@@ -26,10 +26,10 @@ void	show_player(t_cubed *cubed, t_our_img *player, t_canvas *map)
 	int		pos_y;
 
 	mlx = cubed->mlx;
-	pos_x = map->pos_x + ((int)cubed->player->pos_x / map->scale) \
-						- (player->w / 2);
-	pos_y = map->pos_y + ((int)cubed->player->pos_y / map->scale) \
-						- (player->h / 2);
+	pos_x = map->pos_x + (cubed->player->pos_x / map->scale) \
+						- player->w / 2;
+	pos_y = map->pos_y + (cubed->player->pos_y / map->scale) \
+						- player->h / 2;
 	mlx_put_image_to_window(mlx->mlx_ptr, mlx->win_ptr, \
 	player->img, pos_x, pos_y);
 	pos_x = map->pos_x + ((int)cubed->player->dir_x_pos / map->scale \
